feat: add bsp sources to target board

+++ firmware/bsp/STM32F429ZIT/src/include/bsp_display_bus.hpp
@@ -1,13 +1,14 @@
 #ifndef BSP_I2C_HPP_
 #define BSP_I2C_HPP_
 
+#include <cstddef>
 #include <cstdint>
 
 namespace bsp::displayBus {
 
-void write(uint8_t address, uint8_t memoryAddress, const uint8_t *data, uint16_t length);
+void write(uint8_t address, uint8_t memoryAddress, const uint8_t *data, size_t length);
 
-void read(uint8_t address, uint8_t memoryAddress, uint8_t *data, uint16_t length);
+void read(uint8_t address, uint8_t memoryAddress, uint8_t *data, size_t length);
 
 } // namespace bsp::displayBus
